- some changes to traits & weapons

+++ app/src/main/java/com/wekomodo/huntshowdownwiki/data/model/firebase/items/subItems/weapons/Change.kt
@@ -1,7 +1,7 @@
 package com.wekomodo.huntshowdownwiki.data.model.firebase.items.subItems.weapons
 
 data class Change(
-    val from: Int,
+    val from: Double,
     val key: String,
     val label: String,
     val to: Double
